// templates/box/.nango/schema.ts
// ---------------------------
// This file was generated by Nango (v0.63.0)
// You can version this file
// ---------------------------

export interface SyncMetadata_box_users {
};

export interface User {
  id: string;
  email: string;
  firstName: string;
  lastName: string;
};

export interface SyncMetadata_box_folders {
};

export interface Folder {
  id: string;
  name: string;
  modified_at: string;
  url: string | null;
};

export interface SyncMetadata_box_files {
  files: string[];
  folders: string[];
};

export interface BoxDocument {
  id: string;
  name: string;
  download_url: string;
  modified_at: string;
};

export interface ActionInput_box_createuser {
  firstName: string;
  lastName: string;
  email: string;
  address: string;
  can_see_managed_users: boolean;
  external_app_user_id: string;
  is_exempt_from_device_limits: boolean;
  is_exempt_from_login_verification: boolean;
  is_external_collab_restricted: boolean;
  is_platform_access_only: boolean;
  is_sync_enabled: boolean;
  job_title: string;
  language: string;
  phone: string;
  role: 'coadmin' | 'user';
  space_amount: number;
  status: 'active' | 'inactive' | 'cannot_delete_edit' | 'cannot_delete_edit_upload';
  timezone: string;
  tracking_codes: ('tracking_code' | string | string)[];
};

export interface ActionOutput_box_createuser {
  id: string;
  email: string;
  firstName: string;
  lastName: string;
};

export interface ActionInput_box_deleteuser {
  id: string;
  force: boolean;
  notify: boolean;
};

export interface ActionOutput_box_deleteuser {
  success: boolean;
};

export interface ActionInput_box_fetchfile {
  id: string;
};

/** @deprecated It is recommended to use a Model */
export type ActionOutput_box_fetchfile = string

export interface ActionInput_box_foldercontent {
  id: string;
  marker: string;
};

export interface ActionOutput_box_foldercontent {
  files: (string | string | string | string)[];
  folders: (string | string | string | string | null)[];
  next_marker: string;
};